Move login_manager to init to avoid circular imports
